Revert "Fix Col Peripheral Schematic Errors (#386)"
This reverts commit dcd3a332775c58250c864026bd3539b7b3683734.

+++ sramgen/src/layout/sram.rs
@@ -277,7 +277,7 @@ pub fn draw_sram(lib: &mut PdkLib, params: &SramParams) -> Result<PhysicalDesign
         lib,
         &ColInvArrayParams {
             name: "col_inv_array".to_string(),
-            width: cols / mux_ratio,
+            width: cols / mux_ratio as usize,
             mux_ratio,
             instance_params: ColInvParams {
                 name: "col_inv".to_string(),
